new key handler and func pointers in obj to a normal and intersection fix warning in libft (with flag -Weverything)

+++ libraries/libft/ft_atod.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-static double	after_dot(const char *str, int i)
+static double	after_dot(const char *str, size_t i)
 {
 	double	result;
