+++ libraries/libft/ft_atoi.c
@@ -14,9 +14,9 @@
 
 int		ft_atoi(const char *str)
 {
-	int		sign;
-	size_t	i;
-	size_t	result;
+	ssize_t		sign;
+	size_t		i;
+	size_t		result;
 
 	i = 0;
 	ft_skip_white_spases(str, &i);
@@ -27,11 +27,11 @@ int		ft_atoi(const char *str)
 	if (str[i] == '-' || str[i] == '+')
 		i++;
 	while (ft_isdigit(str[i]))
-		result = result * 10 + (int)str[i++] - '0';
+		result = result * 10 + (size_t)str[i++] - '0';
 	if ((result > 9223372036854775807 && sign == 1)
 	|| (ft_digits_len(str) >= 19 && sign == 1))
 		return (-1);
 	if (result > 9223372036854775807 && sign == -1)
 		return (0);
-	return ((int)(result * sign));
+	return ((int)((ssize_t)result * sign));
 }
